Fixed ctx_thread BUG; Fixed bubble caller BUG
ctx_thread.c BUG:
path_store->num_of_paths was going out of whack because we were clearing the
hash table every time we loaded a new colour. This is not a good idea if you
want to construct a path file sharing these colours.

bubble_caller.c BUG:
We were not reversing the supernode first node in walk_supernode_end.

Minor:
* added db_node_match
* improved graph_walker_hash
* merged path_store_thread_safe.c -> graph_paths.c
* wrote graph_paths.c:graph_paths_check_counts() to ensure PathStore + dBGraph
  are in sync

+++ src/kmer/cmd.c
@@ -264,7 +264,8 @@ size_t cmd_get_kmers_in_hash(const CmdArgs *args, size_t extra_bits,
   else if(use_mem_limit)
     graph_mem = hash_table_mem_limit(args->mem_to_use, extra_bits, &kmers_in_hash);
   else if(min_num_kmers > 0)
-    graph_mem = hash_table_mem(min_num_kmers/IDEAL_OCCUPANCY, extra_bits, &kmers_in_hash);
+    graph_mem = hash_table_mem((size_t)(min_num_kmers/IDEAL_OCCUPANCY),
+                               extra_bits, &kmers_in_hash);
   else
     graph_mem = hash_table_mem(1024, extra_bits, &kmers_in_hash);
   // ^ 1024 is a very small default hash table capacity
